requirement.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
pillow
djangorestframework
djangorestframework-simplejwt

